@@ -25,12 +25,13 @@
  * @packageDocumentation
  */
 import { fakeHash, seededRandom } from './fakeHash';
-import { Fraction } from './fractions/bigFraction.js';
-import mimcHash, { modPBigInt, modPBigIntNative } from './mimc';
+import { Fraction } from './fractions/bigFraction';
+import mimcHash, { mimcSponge, modPBigInt, modPBigIntNative } from './mimc';
 import { getRandomGradientAt, IntegerVector, MAX_PERLIN_VALUE, perlin, rand } from './perlin';
 
 export {
   mimcHash,
+  mimcSponge,
   IntegerVector,
   perlin,
   rand,
