@@ -1,15 +1,25 @@
-import { CanvasCoords, GameViewport, RGBAVec, WorldCoords } from '@darkforest_eth/types';
+import {
+  CanvasCoords,
+  CircleRendererType,
+  GameViewport,
+  RendererType,
+  RGBAVec,
+  WorldCoords,
+} from '@darkforest_eth/types';
 import { engineConsts } from '../EngineConsts';
 import { EngineUtils } from '../EngineUtils';
 import { CIRCLE_PROGRAM_DEFINITION } from '../Programs/CircleProgram';
 import { GameGLManager } from '../WebGL/GameGLManager';
 import { GenericRenderer } from '../WebGL/GenericRenderer';
 
-export class CircleRenderer extends GenericRenderer<typeof CIRCLE_PROGRAM_DEFINITION> {
+export class CircleRenderer
+  extends GenericRenderer<typeof CIRCLE_PROGRAM_DEFINITION>
+  implements CircleRendererType
+{
   quadBuffer: number[];
 
   viewport: GameViewport;
-
+  rendererType = RendererType.Circle;
   constructor(manager: GameGLManager) {
     super(manager, CIRCLE_PROGRAM_DEFINITION);
     this.viewport = manager.renderer.getViewport();
