@@ -1,9 +1,22 @@
-import { CanvasCoords, DrawMode, RenderZIndex, RGBAVec, WorldCoords } from '@darkforest_eth/types';
+import {
+  CanvasCoords,
+  DrawMode,
+  LineRendererType,
+  RendererType,
+  RenderZIndex,
+  RGBAVec,
+  WorldCoords,
+} from '@darkforest_eth/types';
 import { LINE_PROGRAM_DEFINITION } from '../Programs/LineProgram';
 import { GameGLManager } from '../WebGL/GameGLManager';
 import { GenericRenderer } from '../WebGL/GenericRenderer';
 
-export class LineRenderer extends GenericRenderer<typeof LINE_PROGRAM_DEFINITION, GameGLManager> {
+export class LineRenderer
+  extends GenericRenderer<typeof LINE_PROGRAM_DEFINITION, GameGLManager>
+  implements LineRendererType
+{
+  rendererType = RendererType.Line;
+
   constructor(glManager: GameGLManager) {
     super(glManager, LINE_PROGRAM_DEFINITION);
   }
