@@ -166,6 +166,32 @@ export type BiomebaseSnarkContractCallArgs = [
 ];
 
 /**
+ * Shape of a javascript object that must be passed into snarkJS `fullProve`
+ * proof generation function for `whitelist` circuit
+ */
+export interface WhitelistSnarkInput {
+  key: string;
+  recipient: string;
+}
+
+/**
+ * Shape of the args for `whitelistRegister` DarkForest contract call.
+ */
+export type WhitelistSnarkContractCallArgs = [
+  [string, string], // proofA
+  [
+    // proofB
+    [string, string],
+    [string, string]
+  ],
+  [string, string], // proofC
+  [
+    string, // hashed whitelist key
+    string // recipient address
+  ]
+];
+
+/**
  * Type representing the shape of args that are passed into DarkForest
  * functions that require zkSNARK verification.
  */
@@ -173,7 +199,8 @@ export type ContractCallArgs =
   | RevealSnarkContractCallArgs
   | InitSnarkContractCallArgs
   | MoveSnarkContractCallArgs
-  | BiomebaseSnarkContractCallArgs;
+  | BiomebaseSnarkContractCallArgs
+  | WhitelistSnarkContractCallArgs;
 
 /**
  * A zkSNARK proof (without signals) generated by snarkJS `fullProve`
@@ -276,3 +303,13 @@ export const biomebaseSnarkWasmPath = require.resolve('./biomebase.wasm');
  * @hidden
  */
 export const biomebaseSnarkZkeyPath = require.resolve('./biomebase.zkey');
+
+/**
+ * @hidden
+ */
+export const whitelistSnarkWasmPath = require.resolve('./whitelist.wasm');
+
+/**
+ * @hidden
+ */
+export const whitelistSnarkZkeyPath = require.resolve('./whitelist.zkey');
