@@ -1,7 +1,7 @@
 import type { Biome, SpaceType } from './game_types';
 import type { ArtifactId, EthAddress, LocationId } from './identifier';
 import type { PlanetMessage } from './planetmessage';
-import { TransactionCollection } from './transaction';
+import type { TransactionCollection } from './transaction';
 import type { Upgrade, UpgradeState } from './upgrade';
 import type { Abstract } from './utility';
 import type { WorldLocation } from './world';
