@@ -1,50 +1,45 @@
 import type { Contract } from 'ethers';
+import type { LiteralUnion } from 'type-fest';
 import type { ArtifactId, EthAddress, LocationId } from './identifier';
 import type { WorldLocation } from './world';
 
 /**
  * @hidden
  */
-//eslint-disable-next-line @projectsophon/typescript-enum/no-enum
-export const enum ContractMethodName {
-  REVEAL_LOCATION = 'revealLocation',
-  INIT = 'initializePlayer',
-  MOVE = 'move',
-  UPGRADE = 'upgradePlanet',
-  BUY_HAT = 'buyHat',
-  PLANET_TRANSFER = 'transferPlanet',
-  FIND_ARTIFACT = 'findArtifact',
-  PROSPECT_PLANET = 'prospectPlanet',
-  DEPOSIT_ARTIFACT = 'depositArtifact',
-  WITHDRAW_ARTIFACT = 'withdrawArtifact',
-  ACTIVATE_ARTIFACT = 'activateArtifact',
-  DEACTIVATE_ARTIFACT = 'deactivateArtifact',
-  WITHDRAW_SILVER = 'withdrawSilver',
-  USE_KEY = 'useKey',
-  ADD_KEYS = 'addKeys',
-  GET_SHIPS = 'giveSpaceShips',
-  CREATE_LOBBY = 'createLobby',
-  INVADE_PLANET = 'invadePlanet',
-  INVADE_TARGET_PLANET = 'invadeTargetPlanet',
-  CAPTURE_PLANET = 'capturePlanet',
-  CLAIM_VICTORY = "claimTargetPlanetVictory",
-  CREATE_PLANET = "createPlanet",
-  CREATE_ARENA_PLANET = "createArenaPlanet",
-}
+export type ContractMethodName =
+  | 'revealLocation'
+  | 'initializePlayer'
+  | 'move'
+  | 'upgradePlanet'
+  | 'buyHat'
+  | 'transferPlanet'
+  | 'findArtifact'
+  | 'prospectPlanet'
+  | 'depositArtifact'
+  | 'withdrawArtifact'
+  | 'activateArtifact'
+  | 'deactivateArtifact'
+  | 'withdrawSilver'
+  | 'useKey'
+  | 'adminUseKey'
+  | 'addKeys'
+  | 'giveSpaceShips'
+  | 'createLobby'
+  | 'invadePlanet'
+  | 'capturePlanet'
+  | "claimTargetPlanetVictory"
+  | "createPlanet"
+  | "createArenaPlanet";
 
-/**
- * @hidden
- */
-//eslint-disable-next-line @projectsophon/typescript-enum/no-enum
-export const enum EthTxStatus {
-  Init = 'Init',
-  Processing = 'Processing',
-  Prioritized = 'Prioritized',
-  Submit = 'Submit',
-  Confirm = 'Confirm',
-  Fail = 'Fail',
-  Cancel = 'Cancel',
-}
+
+export type EthTxStatus =
+  | 'Init'
+  | 'Processing'
+  | 'Prioritized'
+  | 'Submit'
+  | 'Confirm'
+  | 'Fail'
+  | 'Cancel';
 
 /**
  * The intent of this type is to represent a transaction that will occur on the blockchain in a way
@@ -53,7 +48,7 @@ export const enum EthTxStatus {
  */
 export type TxIntent = {
   contract: Contract;
-  methodName: ContractMethodName;
+  methodName: LiteralUnion<ContractMethodName, string>;
   args: Promise<unknown[]>;
 };
 
@@ -61,7 +56,7 @@ export type TxIntent = {
  * @hidden
  */
 export type UnconfirmedInit = TxIntent & {
-  methodName: ContractMethodName.INIT;
+  methodName: 'initializePlayer';
   locationId: LocationId;
   location: WorldLocation;
 };
@@ -70,7 +65,7 @@ export type UnconfirmedInit = TxIntent & {
  * @hidden
  */
 export type UnconfirmedMove = TxIntent & {
-  methodName: ContractMethodName.MOVE;
+  methodName: 'move';
   from: LocationId;
   to: LocationId;
   forces: number;
@@ -83,7 +78,7 @@ export type UnconfirmedMove = TxIntent & {
  * @hidden
  */
 export type UnconfirmedFindArtifact = TxIntent & {
-  methodName: ContractMethodName.FIND_ARTIFACT;
+  methodName: 'findArtifact';
   planetId: LocationId;
 };
 
@@ -91,7 +86,7 @@ export type UnconfirmedFindArtifact = TxIntent & {
  * @hidden
  */
 export type UnconfirmedProspectPlanet = TxIntent & {
-  methodName: ContractMethodName.PROSPECT_PLANET;
+  methodName: 'prospectPlanet';
   planetId: LocationId;
 };
 
@@ -99,7 +94,7 @@ export type UnconfirmedProspectPlanet = TxIntent & {
  * @hidden
  */
 export type UnconfirmedPlanetTransfer = TxIntent & {
-  methodName: ContractMethodName.PLANET_TRANSFER;
+  methodName: 'transferPlanet';
   planetId: LocationId;
   newOwner: EthAddress;
 };
@@ -108,7 +103,7 @@ export type UnconfirmedPlanetTransfer = TxIntent & {
  * @hidden
  */
 export type UnconfirmedUpgrade = TxIntent & {
-  methodName: ContractMethodName.UPGRADE;
+  methodName: 'upgradePlanet';
   locationId: LocationId;
   upgradeBranch: number; // 0, 1, or 2
 };
@@ -117,7 +112,7 @@ export type UnconfirmedUpgrade = TxIntent & {
  * @hidden
  */
 export type UnconfirmedBuyHat = TxIntent & {
-  methodName: ContractMethodName.BUY_HAT;
+  methodName: 'buyHat';
   locationId: LocationId;
 };
 
@@ -125,7 +120,7 @@ export type UnconfirmedBuyHat = TxIntent & {
  * @hidden
  */
 export type UnconfirmedDepositArtifact = TxIntent & {
-  methodName: ContractMethodName.DEPOSIT_ARTIFACT;
+  methodName: 'depositArtifact';
   locationId: LocationId;
   artifactId: ArtifactId;
 };
@@ -134,7 +129,7 @@ export type UnconfirmedDepositArtifact = TxIntent & {
  * @hidden
  */
 export type UnconfirmedWithdrawArtifact = TxIntent & {
-  methodName: ContractMethodName.WITHDRAW_ARTIFACT;
+  methodName: 'withdrawArtifact';
   locationId: LocationId;
   artifactId: ArtifactId;
 };
@@ -143,6 +138,7 @@ export type UnconfirmedWithdrawArtifact = TxIntent & {
  * @hidden
  */
 export type UnconfirmedActivateArtifact = TxIntent & {
+  methodName: 'activateArtifact';
   locationId: LocationId;
   artifactId: ArtifactId;
   wormholeTo?: LocationId;
@@ -152,6 +148,7 @@ export type UnconfirmedActivateArtifact = TxIntent & {
  * @hidden
  */
 export type UnconfirmedDeactivateArtifact = TxIntent & {
+  methodName: 'deactivateArtifact';
   locationId: LocationId;
   artifactId: ArtifactId;
 };
@@ -160,7 +157,7 @@ export type UnconfirmedDeactivateArtifact = TxIntent & {
  * @hidden
  */
 export type UnconfirmedWithdrawSilver = TxIntent & {
-  methodName: ContractMethodName.WITHDRAW_SILVER;
+  methodName: 'withdrawSilver';
   locationId: LocationId;
   amount: number;
 };
@@ -169,7 +166,7 @@ export type UnconfirmedWithdrawSilver = TxIntent & {
  * @hidden
  */
 export type UnconfirmedReveal = TxIntent & {
-  methodName: ContractMethodName.REVEAL_LOCATION;
+  methodName: 'revealLocation';
   locationId: LocationId;
   location: WorldLocation;
 };
@@ -178,21 +175,28 @@ export type UnconfirmedReveal = TxIntent & {
  * @hidden
  */
 export type UnconfirmedAddKeys = TxIntent & {
-  methodName: ContractMethodName.ADD_KEYS;
+  methodName: 'addKeys';
 };
 
 /**
  * @hidden
  */
 export type UnconfirmedUseKey = TxIntent & {
-  methodName: ContractMethodName.USE_KEY;
+  methodName: 'useKey';
+};
+
+/**
+ * @hidden
+ */
+export type UnconfirmedAdminUseKey = TxIntent & {
+  methodName: 'adminUseKey';
 };
 
 /**
  * @hidden
  */
 export type UnconfirmedGetShips = TxIntent & {
-  methodName: ContractMethodName.GET_SHIPS;
+  methodName: 'giveSpaceShips';
   locationId: LocationId;
 };
 
@@ -200,42 +204,34 @@ export type UnconfirmedGetShips = TxIntent & {
  * @hidden
  */
 export type UnconfirmedCreateLobby = TxIntent & {
-  methodName: ContractMethodName.CREATE_LOBBY;
+  methodName: 'createLobby';
 };
 
 /**
  * @hidden
  */
 export type UnconfirmedInvadePlanet = TxIntent & {
-  methodName: ContractMethodName.INVADE_PLANET;
+  methodName: 'invadePlanet';
   locationId: LocationId;
-};
-
-/*
-* @hidden
-*/
-export declare type UnconfirmedInvadeTargetPlanet = TxIntent & {
-   methodName: ContractMethodName.INVADE_TARGET_PLANET;
-   locationId: LocationId;
 };
 
 /**
  * @hidden
  */
 export type UnconfirmedCapturePlanet = TxIntent & {
-  methodName: ContractMethodName.CAPTURE_PLANET;
+  methodName: 'capturePlanet';
   locationId: LocationId;
 };
 /**
  * @hidden
  */
  export declare type UnconfirmedClaimVictory = TxIntent & {
-  methodName: ContractMethodName.CLAIM_VICTORY;
+  methodName: 'claimVictory';
   locationId: LocationId;
 };
 /**
  * @hidden
  */
  export declare type UnconfirmedCreateArenaPlanet = TxIntent & {
-  methodName: ContractMethodName.CREATE_ARENA_PLANET;
+  methodName: 'createArenaPlanet';
 };
